requirements.txt
Flask==2.2.5
PyJWT==2.8.0
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.5.1
eventlet==0.33.3
python-dotenv==1.0.1
psycopg2-binary==2.9.7
pytz
line-bot-sdk==3.5.0
Pillow>=9.0.0
requests>=2.28.0
flask-login==0.6.2
bcrypt==4.0.1
